requirements.txt
# Core dependencies
numpy>=1.18.0
requests>=2.31.0

# AWS deployment
boto3>=1.34.0

# Epic FHIR OAuth
PyJWT>=2.8.0
cryptography>=41.0.0

# ML (optional - for prediction)
scikit-learn>=1.4.0
xgboost>=1.7.0

# Kafka (optional - for sensor pipeline)
confluent-kafka>=1.3.0

# Visualization (optional - for examples)
matplotlib>=3.2.0

# Development
ipython>=7.14.0
line-profiler>=3.0.0
